Remove obsolete higher order components
